requirements.txt
# Core dependencies
torch>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
scipy>=1.11.0

# 多模态特征提取（必需）
torchvision>=0.15.0      # ResNet图像特征
transformers>=4.30.0     # BERT文本特征
pillow>=9.0.0            # 图像处理
requests>=2.28.0         # 图片下载

# 模型训练
einops>=0.7.0
tensorboard>=2.13.0
tqdm>=4.65.0
pyyaml>=6.0.0

# 可视化
matplotlib>=3.7.0
seaborn>=0.12.0

# Optional: for quantum computing support
# qiskit>=0.45.0
# pennylane>=0.32.0
